Fix clang-tidy warnings found by make check
- Fix use-after-free in drm_property GETBLOB and drm_vblank WAIT_VBLANK
  (read fields before releasing the object)
- Fix uninitialized actual_size in gzip_decompress
- Make explicit uint64 multiplication in TIMER_HZ-based net macros
  (ARP/IPv4/IPv6/NDP/TCP timeouts) to avoid 32-bit overflow
- Fix potential 32-bit overflow doubling TCP persist interval
- Sort includes with clang-format across affected files

+++ net/ipv4/arp.c
@@ -12,14 +12,14 @@
 #include <kernel/printk.h>
 #include <kernel/timer/timer.h>
 #include <libs/std/string.h>
-#include <net/ipv4/arp.h>
 #include <net/core/endian.h>
 #include <net/core/ethernet.h>
+#include <net/ipv4/arp.h>
 #include <net/ipv4/ipv4.h>
 #include <proc/sched.h>
 
 #define ARP_PACKET_LEN  28U
-#define ARP_TTL_TICKS   (60U * TIMER_HZ)
+#define ARP_TTL_TICKS   ((uint64_t)60U * TIMER_HZ)
 #define ARP_RETRY_TICKS TIMER_HZ
 #define ARP_MAX_RETRIES 3U
