+++ net/ipv4/ipv4.c
@@ -13,15 +13,15 @@
 #include <kernel/timer/timer.h>
 #include <libs/std/string.h>
 #include <mem/alloc.h>
-#include <net/ipv4/arp.h>
 #include <net/core/endian.h>
+#include <net/ipv4/arp.h>
 #include <net/ipv4/icmp.h>
 #include <net/ipv4/ipv4.h>
 #include <net/transport/tcp.h>
 #include <net/transport/udp.h>
 #include <proc/sched.h>
 
-#define IPV4_REASSEMBLY_TIMEOUT_TICKS (30U * TIMER_HZ)
+#define IPV4_REASSEMBLY_TIMEOUT_TICKS ((uint64_t)30U * TIMER_HZ)
 #define IPV4_MAX_HEADER               60U
 #define IPV4_MAX_PAYLOAD              (UINT16_MAX - IPV4_HEADER_MIN)
 #define IPV4_BITMAP_SIZE              ((IPV4_MAX_PAYLOAD + 7U) / 8U)
